Fix location of netapp module_utils unit tests.
